src/pythonservices/requirements.txt
fastapi
uvicorn[standard]
numpy
pandas
scikit-learn
python-dotenv
supabase
pydantic
gunicorn
python-multipart
